requirements.txt
langchain>=0.2.11
langchain-core>=0.2.38
langchain-community>=0.2.10
langchain-ollama>=0.1.0
pydantic>=2.7.1
typer>=0.12.3
uvicorn>=0.29.0
fastapi>=0.111.0
httpx>=0.27.0
python-dotenv>=1.0.1
json5>=0.9.25
